app/requirements.txt
django~=3.1.3
django-crispy-forms~=1.10.0
django-bootstrap4~=2.3.1
django-bootstrap-datepicker-plus~=3.0.5
django-money~=1.2
gunicorn~=20.0.4
django-mathfilters~=1.0.0
psycopg2-binary==2.8.6